// bao-source/bao-shared/src/generated/flatbuffers/baohaus/realtime/v1/battery-status.ts
// automatically generated by the FlatBuffers compiler, do not modify
import { Builder } from '@baohaus/flatbuf-bao/builder';
import type { IGeneratedObject, Offset } from '@baohaus/flatbuf-bao/builder';
import { ByteBuffer } from '@baohaus/flatbuf-bao/reader';
import type { IUnpackableObject } from '@baohaus/flatbuf-bao/reader';
export class BatteryStatus implements IUnpackableObject<BatteryStatusT> {
  bb: ByteBuffer|null = null;
  bb_pos = 0;
  __init(i:number, bb:ByteBuffer):BatteryStatus {
  this.bb_pos = i;
  this.bb = bb;
  return this;
}

level():number {
  return this.bb!.readFloat32(this.bb_pos);
}

voltage():number {
  return this.bb!.readFloat32(this.bb_pos + 4);
}

current():number {
  return this.bb!.readFloat32(this.bb_pos + 8);
}

temperature():number {
  return this.bb!.readFloat32(this.bb_pos + 12);
}

remainingPercent():number {
  return this.bb!.readFloat32(this.bb_pos + 16);
}

static sizeOf():number {
  return 20;
}

static createBatteryStatus(builder:Builder, level: number, voltage: number, current: number, temperature: number, remainingPercent: number):Offset {
  builder.prep(4, 20);
  builder.writeFloat32(remainingPercent);
  builder.writeFloat32(temperature);
  builder.writeFloat32(current);
  builder.writeFloat32(voltage);
  builder.writeFloat32(level);
  return builder.offset();
}


unpack(): BatteryStatusT {
  return new BatteryStatusT(
    this.level(),
    this.voltage(),
    this.current(),
    this.temperature(),
    this.remainingPercent()
  );
}


unpackTo(_o: BatteryStatusT): void {
  _o.level = this.level();
  _o.voltage = this.voltage();
  _o.current = this.current();
  _o.temperature = this.temperature();
  _o.remainingPercent = this.remainingPercent();
}
}

export class BatteryStatusT implements IGeneratedObject {
constructor(
  public level: number = 0.0,
  public voltage: number = 0.0,
  public current: number = 0.0,
  public temperature: number = 0.0,
  public remainingPercent: number = 0.0
){}


pack(builder:Builder): Offset {
  return BatteryStatus.createBatteryStatus(builder,
    this.level,
    this.voltage,
    this.current,
    this.temperature,
    this.remainingPercent
  );
}
}